pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "vmware-avi"
version = "1.4.3"
description = "AVI (NSX Advanced Load Balancer) management and AKO Kubernetes operations tool"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [{ name = "Wei Zhou / 周崴" }]
dependencies = [
    "avisdk>=22.1,<31.0",
    "kubernetes>=28.0,<33.0",
    "pyaml>=24.0,<27.0",
    "typer>=0.12,<1.0",
    "rich>=13.0,<15.0",
    "httpx>=0.27,<1.0",
    "python-dotenv>=1.0,<2.0",
    "mcp[cli]>=1.0,<2.0",
    "vmware-policy>=1.0.0,<2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0,<10.0",
    "pytest-cov>=5.0,<8.0",
    "ruff>=0.5,<1.0",
]

[project.scripts]
vmware-avi = "vmware_avi.cli:app"
vmware-avi-mcp = "mcp_server.server:main"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "unit: Unit tests",
    "integration: Integration tests (require AVI Controller)",
]
